+++ src/main/java/dev/krona/urbex/setup/Registration.java
@@ -6,11 +6,7 @@ import dev.krona.urbex.worldgen.CityFeature;
 import dev.krona.urbex.worldgen.SphereFeature;
 import net.minecraft.core.Registry;
 import net.minecraft.core.registries.BuiltInRegistries;
-import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.Identifier;
-import net.minecraft.resources.ResourceKey;
-import net.minecraft.world.level.Level;
-import net.minecraft.world.level.dimension.DimensionType;
 import net.minecraft.world.level.levelgen.feature.Feature;
 
 
@@ -33,9 +29,4 @@ public class Registration {
         sphereFeature = Registry.register(BuiltInRegistries.FEATURE,
                 Identifier.fromNamespaceAndPath(Urbex.MODID, "spheres"), new SphereFeature());
     }
-
-    public static final Identifier CITY_ID = Identifier.fromNamespaceAndPath(Urbex.MODID, "city");
-
-    public static final ResourceKey<DimensionType> DIMENSION_TYPE = ResourceKey.create(Registries.DIMENSION_TYPE, CITY_ID);
-    public static final ResourceKey<Level> DIMENSION = ResourceKey.create(Registries.DIMENSION, CITY_ID);
 }